Some update in the aplication
Modifications finalizing the upload process, except that it isn't processed by open game benchmarks properly and because of that just the login process is successful, the rest don't produce any error but isn't registered in the site. Also the login password don't appear anymore while inserting it in the console.